rpmsg: virtio_rpmsg_bus: Use workqueue for registering/unregistering devices
When virtio_rpmsg_bus rpmsg name service registers a new rpmsg device,
it is doing so from a receive message callback. This means that the
message receive queue processing in rpmsg_recv_done() is blocked until
the device registration is done.

If a rpmsg driver wants to receive a message during probe, this causes
a deadlock because the receive queue is currently blocked until the
rpmsg driver probe function returns.

This defers the adding and removing of rpmsg devices to a workqueue to
avoid this potential deadlock.

Signed-off-by: David Lechner <david@lechnology.com>